Implemented transfer method in service and controller.

+++ src/main/java/com/example/banksystemapi/dto/TransactionDto.java
@@ -1,6 +1,7 @@
 package com.example.banksystemapi.dto;
 
 
+import jakarta.persistence.criteria.CriteriaBuilder;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,8 +12,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TransactionDto {
-    private Long originatingAccountId;
-    private Long resultingAccountId;
+    private Integer originatingAccountId;
+    private Integer resultingAccountId;
     private double amount;
     private String description;
 }
